+++ lib/src/repositories/addresses/addresses.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:elrond_gateway/src/response/addresses/get_address/get_address.dart';
+import 'package:elrond_gateway/src/response/addresses/get_address_nonce/get_address_nonce.dart';
 import 'package:retrofit/retrofit.dart';
 
 part 'addresses.g.dart';
@@ -10,4 +11,7 @@ abstract class Addresses {
 
   @GET('/address/{bech32Address}')
   Future<GetAddressResponse> getAddress();
+
+  @GET('/address/{bech32Address}/nonce')
+  Future<GetAddressNonceResponse> getAddressNonce();
 }
